add callchain-dubstep.h and ccx tests
this allows calling with the function in brackets:

    callchain2(x)
              [foo]()
              [bar](y,z)
              [baz]()
              ()

    -->  baz(bar(foo(x),y,z))

callchain2 uses alternating helper classes, one for storing a value,
and the other for keeping references to the value and a function.

Because the intermediate values (if returned as rvalues) are stored by
value in the value class, and because the function is passed in directly
in the top-level expression (rather than computed inside a callchain
member function), it seems fine to keep references in the function+value
class, even to rvalues, as the objects they refer to will not be
destroyed before they need to be used.  (They will all live for the
lifetime of the whole callchain2 expression.)

As a result (of always keeping references in the function+value helper
class), the "semantics" appear to be exactly the same as the
single-helper-class version -- exactly the same constructors &
destructors are called for intermediate objects.  And the extra
reference holding of the second class all gets optimized out anyway.